move Inventory to new inventory.py file
